Perfected

+++ WielkaApkaLegendarna/Assets/Scripts/SaveToJSON.cs
@@ -2,10 +2,13 @@ using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class SaveToJSON : MonoBehaviour
 {
     [SerializeField] Builder builder;
+    [SerializeField] bool overwiter;
+    [SerializeField] Image overwriteWarning;
 
     public void SaveToFile()
     {
@@ -22,7 +25,23 @@ public class SaveToJSON : MonoBehaviour
         {
             Directory.CreateDirectory(Application.dataPath + "/Characters");
         }
-        File.WriteAllText(Application.dataPath + "/Characters" + "/" + character.imie + ".json", json);
+
+        if (File.Exists(Application.dataPath + "/Characters" + "/" + character.imie + ".json") && overwiter)
+        {
+            File.WriteAllText(Application.dataPath + "/Characters" + "/" + character.imie + ".json", json);
+            overwriteWarning.gameObject.SetActive(false);
+        }
+
+        if (File.Exists(Application.dataPath + "/Characters" + "/" + character.imie + ".json") && !overwiter)
+        {
+            overwriteWarning.gameObject.SetActive(true);
+        }      
+
+        if (!File.Exists(Application.dataPath + "/Characters" + "/" + character.imie + ".json"))
+        {
+            File.WriteAllText(Application.dataPath + "/Characters" + "/" + character.imie + ".json", json);
+        }
+
 
     }
 }
